+++ kitty/state.h
@@ -80,6 +80,7 @@ typedef struct {
     struct {
         double outer, inner;
     } tab_bar_margin_height;
+    long macos_menubar_title_max_length;
 } Options;
 
 typedef struct WindowLogoRenderData {